+++ plugins/main/public/components/endpoints-summary/register-agent/containers/register-agent/register-agent.tsx
@@ -4,12 +4,10 @@ import {
   EuiFlexItem,
   EuiPanel,
   EuiTitle,
-  EuiButtonEmpty,
   EuiPage,
   EuiPageBody,
   EuiSpacer,
   EuiProgress,
-  EuiButton,
 } from '@elastic/eui';
 import { WzRequest } from '../../../../../react-services/wz-request';
 import { UI_LOGGER_LEVELS } from '../../../../../../common/constants';
@@ -44,7 +42,6 @@ import {
   nestedResolve,
   savedSearch,
 } from '../../../../../services/resolves';
-import NavigationService from '../../../../../react-services/navigation-service';
 
 export const RegisterAgent = compose(
   withErrorBoundary,
@@ -184,31 +181,15 @@ export const RegisterAgent = compose(
 
   return (
     <div>
-      <EuiPage restrictWidth='1000px' style={{ background: 'transparent' }}>
+      <EuiPage style={{ background: 'transparent' }}>
         <EuiPageBody>
           <EuiFlexGroup>
             <EuiFlexItem>
               <EuiPanel className='register-agent-wizard-container'>
-                <div className='register-agent-wizard-close'>
-                  <EuiButtonEmpty
-                    size='s'
-                    href={NavigationService.getInstance().getUrlForApp(
-                      endpointSummary.id,
-                      {
-                        path: `#${endpointSummary.redirectTo()}`,
-                      },
-                    )}
-                    iconType='cross'
-                  >
-                    Close
-                  </EuiButtonEmpty>
-                </div>
                 <EuiFlexGroup>
                   <EuiFlexItem>
-                    <EuiTitle>
-                      <h2 className='register-agent-wizard-title'>
-                        Deploy new agent
-                      </h2>
+                    <EuiTitle size='s'>
+                      <h1>Deploy new agent</h1>
                     </EuiTitle>
                   </EuiFlexItem>
                 </EuiFlexGroup>
@@ -233,26 +214,6 @@ export const RegisterAgent = compose(
                     />
                   </EuiFlexItem>
                 )}
-                <EuiFlexGroup
-                  justifyContent='flexEnd'
-                  style={{ marginRight: '0.3rem' }}
-                >
-                  <EuiFlexItem grow={false}>
-                    <EuiButton
-                      className='close-button'
-                      fill
-                      color='primary'
-                      href={NavigationService.getInstance().getUrlForApp(
-                        endpointSummary.id,
-                        {
-                          path: `#${endpointSummary.redirectTo()}`,
-                        },
-                      )}
-                    >
-                      Close
-                    </EuiButton>
-                  </EuiFlexItem>
-                </EuiFlexGroup>
               </EuiPanel>
             </EuiFlexItem>
           </EuiFlexGroup>